INFRA-8789 Fix github ssh-auth test (#71)
Dotfiles was using the human readable output of ssh -T to check if
ssh auth was working. This output is not consistent across versions
of osx, so we now use the exit code of ssh -T to check if auth is
working. (ssh returns 1 if auth succeeds, 255 if it fails (and 130
if passphrase is wrong).

I think we saw this problem before when upgrading to big sur
(or some prior version of os x).

Issue: https://khanacademy.atlassian.net/browse/INFRA-8789

Test plan:

Run "make install"
Test it with no ssh keys installed and see if you get prompted to
create one. Then run it with ssh keys installed and verify you do
not get prompted.